+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Prop `skusFilter` to ProductSummaryList.
+
+### Fixed
+- Props passed to ProductSummaryList.
 
 ## [2.51.5] - 2020-02-21
 ### Changed